+++ crates/biome_js_syntax/src/expr_ext.rs
@@ -966,6 +966,13 @@ impl AnyJsExpression {
                 let member = member.as_js_identifier_expression()?.name().ok()?;
                 member.value_token().ok()
             }
+            AnyJsExpression::JsTemplateExpression(node) => {
+                let tag = node.tag()?;
+                let tag = tag.as_js_static_member_expression()?;
+                let member = tag.object().ok()?;
+                let member = member.as_js_identifier_expression()?.name().ok()?;
+                member.value_token().ok()
+            }
             AnyJsExpression::JsIdentifierExpression(node) => node.name().ok()?.value_token().ok(),
             _ => None,
         }
@@ -975,6 +982,13 @@ impl AnyJsExpression {
         match self {
             AnyJsExpression::JsStaticMemberExpression(node) => {
                 let member = node.member().ok()?;
+                let member = member.as_js_name()?;
+                member.value_token().ok()
+            }
+            AnyJsExpression::JsTemplateExpression(node) => {
+                let tag = node.tag()?;
+                let tag = tag.as_js_static_member_expression()?;
+                let member = tag.member().ok()?;
                 let member = member.as_js_name()?;
                 member.value_token().ok()
             }